+++ src/truepyfstlib/__init__.py
@@ -25,7 +25,7 @@ from .common import (
     FstHierType,
 )
 
-__version__ = "0.3.0"
+__version__ = "0.4.0"
 __all__ = [
     "FstReader",
     "FstWriter",